requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
bcrypt
python-multipart
sqlite3
pyjwt
psycopg2-binary
python-jose
python-dotenv